Add plugin and corresponding tests (#3)
This commit adds the actual vite plugin - `dynamicFileSwap`. This plugin allows vite to dynamically swap file contents for specifically tagged imports.

E.g.
```
import { data, isFallback } from "my-custom-import-tag";
```

When added to a vite config file with customized options, this plugin will swap file content for a tagged import from one file to a fallback file and vice versa. This swap is based on the presence of the main file in the working filesystem.