Reorganize code into modules and fix __dirname incompatibility (#6)
This commit reorganizes the single index.js file into separate modules.
This allows for better separation of concerns and more maintainable
tests.

It also fixes a core issue that made the plugin incompatible with vite
build systems, namely that `__dirname` is not accessible within an esm
context. In an earlier attempt locally, I added an isomorphic
`__dirname` polyfill that used `__dirname` in a cjs context, and file
urls for esm, but that was a bit complex and was not compatible with the
esbuild target I was using previously.

I decided to forego this solution and instead use `process.cwd()` as the
base path to resolve the main and fallback file paths. This base path
can be overridden in the passed-in options, but it should be an absolute
base path that can be resolved with the relative main and fallback paths
within the plugin's node module scope.